Put test in correct directory
